requirements.txt
streamlit
ollama
PyPDF2
pandas
requests
arxiv
numpy
scikit-learn
transformers
torch
SQLAlchemy
Werkzeug
flask_wtf
wtforms
langchain
langchain_google_genai
google-generativeai
cachetools
python-dotenv
asyncio
langchain_community
email_validator
faiss-cpu
streamlit-webrtc
gtts
speechrecognition
pydub
whisper
